+++ packages/nuxt/tsdown.config.ts
@@ -7,6 +7,7 @@ export default defineConfig({
 		"src/module.ts",
 		"src/server.ts",
 		"src/client.ts",
+		"src/empty-client-env.ts",
 		"src/standard/index.ts",
 		"src/standard/module.ts",
 		"src/standard/server.ts",
@@ -20,6 +21,6 @@ export default defineConfig({
 	sourcemap: true,
 	deps: {
 		alwaysBundle: ["@repo/log", "@repo/types", "@repo/utils"],
-		neverBundle: ["@nuxt/kit", "@nuxt/schema"],
+		neverBundle: ["@nuxt/kit", "@nuxt/schema", "#arkenv/client-env"],
 	},
 });