+++ packages/nuxt/vitest.config.ts
@@ -19,6 +19,10 @@ export default defineProject({
 				find: /^arkenv$/,
 				replacement: path.resolve(__dirname, "../core/src/index.ts"),
 			},
+			{
+				find: "#arkenv/client-env",
+				replacement: path.resolve(__dirname, "./src/empty-client-env.ts"),
+			},
 		],
 	},
 });